+++ test cases/cmake/28 config overrides/subprojects/cmMod/cmMod.cpp
@@ -1,4 +1,4 @@
-#include "cmMod.hpp"
+#include "include/cmMod.hpp"
 
 using namespace std;
